% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{data}
\name{adj}
\alias{adj}
\title{An adjacent information for example dataset.}
\format{
A data frame with 1180 rows and 2 variables:
\describe{
\item{lab}{labels for spatio-temporal}
\item{adj}{adjacent labels}
}
}
\source{
none
}
\usage{
adj
}
\description{
A dataset is the adjacent information for spatio-temporal in example dataset.
}
\keyword{datasets}
