% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata.R
\docType{data}
\name{exdata}
\alias{exdata}
\title{An example dataset.}
\format{
A data frame with 200 rows and 5 variables:
\describe{
\item{y}{response variables according to binomial distribution}
\item{m}{the numbers of trials}
\item{area}{locations}
\item{year}{time points}
\item{lab}{labels for spatio-temporal}
}
}
\source{
none
}
\usage{
exdata
}
\description{
A dataset is spatio-temporal data for 10 areas and 20 time points.
}
\keyword{datasets}
