% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLlogit.R
\name{GFLlogit}
\alias{GFLlogit}
\title{Coordinate optimization for GFL logistic regression}
\usage{
GFLlogit(
  m,
  y,
  adjCD,
  lambda = NULL,
  lambda.type = NULL,
  weight = NULL,
  conv = "coef",
  progress = FALSE,
  thres = 1e-05
)
}
\arguments{
\item{m}{a vector of the numbers of trials}

\item{y}{a vector of the numbers of successes}

\item{adjCD}{a list of adjacent information}

\item{lambda}{tuning parameter; "NULL" or vector/scalar}

\item{lambda.type}{option when is.null(lambda)=FALSE; value:
lambda is searching points; rate: lam.max*lambda is searching points}

\item{weight}{penalty weights; "NULL" or list}

\item{conv}{"coef" or "pll"; A convergence judgement is based on "conv"}

\item{progress}{If TRUE, progress is displayed}

\item{thres}{a threshold for convergence judgement}
}
\value{
mu.hat: vector of estimates

lam.hat: the optimal tuning parameter

pll: the minimum of the objective function

time: runtime (s)
}
\description{
\code{GFLlogit} This function calculates estimates for GFL logistic regression
via coordinate descent algorithm.
}
\examples{
#GFLlogit(m, y, adjCD)
}
