% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLlogit1.R
\name{GFLlogit1}
\alias{GFLlogit1}
\title{Coordinate optimization for GFL logistic regression}
\usage{
GFLlogit1(m, y, lambda, w, z)
}
\arguments{
\item{m}{natural value}

\item{y}{natural value (y < m)}

\item{lambda}{positive value}

\item{w}{vector of positive values}

\item{z}{vector which are all distinct values}
}
\value{
the minimizer
}
\description{
\code{GFLlogit1} This function is used in the function "GFLlogit"
}
\examples{
#GFLlogit1(m, y, labmda, w, z)
}
